+++ lib/src/apps/app_stores/app_stores.dart
@@ -1,3 +1,4 @@
+export 'huawei_app_gallery_store.dart';
 export 'ios_app_store.dart';
 export 'mac_app_store.dart';
 export 'microsoft_store.dart';